+++ packages/cosmwasm-stargate/src/aminotypes.ts
@@ -8,18 +8,19 @@ import {
   MsgMigrateContract,
   MsgStoreCode,
   MsgUpdateAdmin,
-} from "cosmjs-types/cosmwasm/wasm/v1beta1/tx";
+} from "cosmjs-types/cosmwasm/wasm/v1/tx";
 import Long from "long";
 
 // TODO: implement
 /**
- * @see https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/types.proto#L30-L35
+ * @see https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/types.proto#L36-L41
  */
 type AccessConfig = never;
 
 /**
- * The Amino JSON representation of
- * https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/tx.proto#L28-L40
+ * The Amino JSON representation of [MsgStoreCode].
+ *
+ * [MsgStoreCode]: https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/tx.proto#L28-L39
  */
 export interface AminoMsgStoreCode {
   type: "wasm/MsgStoreCode";
@@ -28,17 +29,14 @@ export interface AminoMsgStoreCode {
     readonly sender: string;
     /** Base64 encoded Wasm */
     readonly wasm_byte_code: string;
-    /** A valid URI reference to the contract's source code. Can be empty. */
-    readonly source: string;
-    /** A docker tag. Can be empty. */
-    readonly builder: string;
     readonly instantiate_permission?: AccessConfig;
   };
 }
 
 /**
- * The Amino JSON representation of
- * https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/tx.proto#L70-L80
+ * The Amino JSON representation of [MsgExecuteContract].
+ *
+ * [MsgExecuteContract]: https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/tx.proto#L73-L86
  */
 export interface AminoMsgExecuteContract {
   type: "wasm/MsgExecuteContract";
@@ -54,8 +52,9 @@ export interface AminoMsgExecuteContract {
 }
 
 /**
- * The Amino JSON representation of
- * https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/tx.proto#L47-L61
+ * The Amino JSON representation of [MsgInstantiateContract].
+ *
+ * [MsgInstantiateContract]: https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/tx.proto#L46-L64
  */
 export interface AminoMsgInstantiateContract {
   type: "wasm/MsgInstantiateContract";
@@ -66,8 +65,8 @@ export interface AminoMsgInstantiateContract {
     readonly code_id: string;
     /** Human-readable label for this contract */
     readonly label: string;
-    /** Init message as JavaScript object */
-    readonly init_msg: any;
+    /** Instantiate message as JavaScript object */
+    readonly msg: any;
     readonly funds: readonly Coin[];
     /** Bech32-encoded admin address */
     readonly admin?: string;
@@ -75,8 +74,9 @@ export interface AminoMsgInstantiateContract {
 }
 
 /**
- * The Amino JSON representation of
- * https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/tx.proto#L88-L98
+ * The Amino JSON representation of [MsgMigrateContract].
+ *
+ * [MsgMigrateContract]: https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/tx.proto#L94-L104
  */
 export interface AminoMsgMigrateContract {
   type: "wasm/MsgMigrateContract";
@@ -93,8 +93,9 @@ export interface AminoMsgMigrateContract {
 }
 
 /**
- * The Amino JSON representation of
- * https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/tx.proto#L107-L115
+ * The Amino JSON representation of [MsgUpdateAdmin].
+ *
+ * [MsgUpdateAdmin]: https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/tx.proto#L113-L121
  */
 export interface AminoMsgUpdateAdmin {
   type: "wasm/MsgUpdateAdmin";
@@ -109,8 +110,9 @@ export interface AminoMsgUpdateAdmin {
 }
 
 /**
- * The Amino JSON representation of
- * https://github.com/CosmWasm/wasmd/blob/v0.16.0-alpha2/proto/cosmwasm/wasm/v1beta1/tx.proto#L120-L126
+ * The Amino JSON representation of [MsgClearAdmin].
+ *
+ * [MsgClearAdmin]: https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/tx.proto#L126-L132
  */
 export interface AminoMsgClearAdmin {
   type: "wasm/MsgClearAdmin";
@@ -123,36 +125,32 @@ export interface AminoMsgClearAdmin {
 }
 
 export const cosmWasmTypes: Record<string, AminoConverter> = {
-  "/cosmwasm.wasm.v1beta1.MsgStoreCode": {
+  "/cosmwasm.wasm.v1.MsgStoreCode": {
     aminoType: "wasm/MsgStoreCode",
-    toAmino: ({ sender, wasmByteCode, source, builder }: MsgStoreCode): AminoMsgStoreCode["value"] => ({
+    toAmino: ({ sender, wasmByteCode }: MsgStoreCode): AminoMsgStoreCode["value"] => ({
       sender: sender,
       wasm_byte_code: toBase64(wasmByteCode),
-      source: source,
-      builder: builder,
     }),
-    fromAmino: ({ sender, wasm_byte_code, source, builder }: AminoMsgStoreCode["value"]): MsgStoreCode => ({
+    fromAmino: ({ sender, wasm_byte_code }: AminoMsgStoreCode["value"]): MsgStoreCode => ({
       sender: sender,
       wasmByteCode: fromBase64(wasm_byte_code),
-      source: source,
-      builder: builder,
       instantiatePermission: undefined,
     }),
   },
-  "/cosmwasm.wasm.v1beta1.MsgInstantiateContract": {
+  "/cosmwasm.wasm.v1.MsgInstantiateContract": {
     aminoType: "wasm/MsgInstantiateContract",
     toAmino: ({
       sender,
       codeId,
       label,
-      initMsg,
+      msg,
       funds,
       admin,
     }: MsgInstantiateContract): AminoMsgInstantiateContract["value"] => ({
       sender: sender,
       code_id: codeId.toString(),
       label: label,
-      init_msg: JSON.parse(fromUtf8(initMsg)),
+      msg: JSON.parse(fromUtf8(msg)),
       funds: funds,
       admin: admin ?? undefined,
     }),
@@ -160,19 +158,19 @@ export const cosmWasmTypes: Record<string, AminoConverter> = {
       sender,
       code_id,
       label,
-      init_msg,
+      msg,
       funds,
       admin,
     }: AminoMsgInstantiateContract["value"]): MsgInstantiateContract => ({
       sender: sender,
       codeId: Long.fromString(code_id),
       label: label,
-      initMsg: toUtf8(JSON.stringify(init_msg)),
+      msg: toUtf8(JSON.stringify(msg)),
       funds: [...funds],
       admin: admin ?? "",
     }),
   },
-  "/cosmwasm.wasm.v1beta1.MsgUpdateAdmin": {
+  "/cosmwasm.wasm.v1.MsgUpdateAdmin": {
     aminoType: "wasm/MsgUpdateAdmin",
     toAmino: ({ sender, newAdmin, contract }: MsgUpdateAdmin): AminoMsgUpdateAdmin["value"] => ({
       sender: sender,
@@ -185,7 +183,7 @@ export const cosmWasmTypes: Record<string, AminoConverter> = {
       contract: contract,
     }),
   },
-  "/cosmwasm.wasm.v1beta1.MsgClearAdmin": {
+  "/cosmwasm.wasm.v1.MsgClearAdmin": {
     aminoType: "wasm/MsgClearAdmin",
     toAmino: ({ sender, contract }: MsgClearAdmin): AminoMsgClearAdmin["value"] => ({
       sender: sender,
@@ -196,7 +194,7 @@ export const cosmWasmTypes: Record<string, AminoConverter> = {
       contract: contract,
     }),
   },
-  "/cosmwasm.wasm.v1beta1.MsgExecuteContract": {
+  "/cosmwasm.wasm.v1.MsgExecuteContract": {
     aminoType: "wasm/MsgExecuteContract",
     toAmino: ({ sender, contract, msg, funds }: MsgExecuteContract): AminoMsgExecuteContract["value"] => ({
       sender: sender,
@@ -211,18 +209,13 @@ export const cosmWasmTypes: Record<string, AminoConverter> = {
       funds: [...funds],
     }),
   },
-  "/cosmwasm.wasm.v1beta1.MsgMigrateContract": {
+  "/cosmwasm.wasm.v1.MsgMigrateContract": {
     aminoType: "wasm/MsgMigrateContract",
-    toAmino: ({
-      sender,
-      contract,
-      codeId,
-      migrateMsg,
-    }: MsgMigrateContract): AminoMsgMigrateContract["value"] => ({
+    toAmino: ({ sender, contract, codeId, msg }: MsgMigrateContract): AminoMsgMigrateContract["value"] => ({
       sender: sender,
       contract: contract,
       code_id: codeId.toString(),
-      msg: JSON.parse(fromUtf8(migrateMsg)),
+      msg: JSON.parse(fromUtf8(msg)),
     }),
     fromAmino: ({
       sender,
@@ -233,7 +226,7 @@ export const cosmWasmTypes: Record<string, AminoConverter> = {
       sender: sender,
       contract: contract,
       codeId: Long.fromString(code_id),
-      migrateMsg: toUtf8(JSON.stringify(msg)),
+      msg: toUtf8(JSON.stringify(msg)),
     }),
   },
 };