+++ packages/cosmwasm-stargate/src/cosmwasmclient.spec.ts
@@ -229,8 +229,6 @@ describe("CosmWasmClient", () => {
       const [first] = result;
       expect(first).toEqual({
         id: deployedHackatom.codeId,
-        source: deployedHackatom.source,
-        builder: deployedHackatom.builder,
         checksum: deployedHackatom.checksum,
         creator: alice.address0,
       });
@@ -245,8 +243,6 @@ describe("CosmWasmClient", () => {
 
       const expectedInfo: Code = {
         id: deployedHackatom.codeId,
-        source: deployedHackatom.source,
-        builder: deployedHackatom.builder,
         checksum: deployedHackatom.checksum,
         creator: alice.address0,
       };
@@ -313,12 +309,12 @@ describe("CosmWasmClient", () => {
     it("works for instance with IBC port ID", async () => {
       pendingWithoutWasmd();
       const client = await CosmWasmClient.connect(wasmd.endpoint);
-      const contract = await client.getContract(deployedIbcReflect.instances[0]);
+      const contract = await client.getContract(deployedIbcReflect.instances[0].address);
       expect(contract).toEqual(
         jasmine.objectContaining({
-          address: deployedIbcReflect.instances[0],
+          address: deployedIbcReflect.instances[0].address,
           codeId: deployedIbcReflect.codeId,
-          ibcPortId: "wasm.wasm1vjecguu37pmd577339wrdp208ddzymku8yy0te",
+          ibcPortId: deployedIbcReflect.instances[0].ibcPortId,
         }),
       );
     });